Split quantity tests into parameter-related tests and model evaluation
